refactor: separate out webpacking for tests
