+++ package.json
@@ -79,6 +79,7 @@
     "build:snippets": "tsc --build ./snippets/tsconfig.json",
     "build:lib": "tsc --build tsconfig.build.json",
     "build:web": "webpack",
+    "build:browserTests": "webpack --config ./test/webpack.config.js",
     "analyze": "run-s build:web --analyze",
     "watch": "run-s build:lib --watch",
     "clean": "rm -rf dist",
@@ -86,7 +87,7 @@
     "prepublish": "run-s clean build",
     "test": "nyc mocha --config=test/.mocharc.json --exit",
     "test:integration": "TS_NODE_PROJECT=tsconfig.build.json nyc mocha ./test/integration/**/*.ts ./test/integration/*.ts",
-    "test:browser": "TS_NODE_PROJECT=tsconfig.build.json nyc mocha ./test/browser/*.ts",
+    "test:browser": "npm run build:browserTests && TS_NODE_PROJECT=tsconfig.build.json nyc mocha ./test/browser/*.ts",
     "test:watch": "TS_NODE_PROJECT=src/tsconfig.json mocha --config=test/.mocharc.json --watch --reporter dot",
     "format": "prettier --write '{src,test}/**/*.ts'",
     "lint": "eslint . --ext .ts --max-warnings 0",
