+++ test/localIntegrationRunner.html
@@ -17,7 +17,7 @@
       mocha.ui('bdd')
     </script>
 
-    <script src="../testCompiledForWeb/index.js"></script>
+    <script src="./testCompiledForWeb/index.js"></script>
 
     <script>
       mocha.run()
